+++ C/Jugglingalgorithm.c
