+++ C/Palindrome.c
