+++ C/Sumofthreestacks.c
