+++ C/gcd.c
