+++ C/min_max_of_an_array.c
