+++ C/string4num.c
