+++ C/swappingwithoutthirdvariable.c
