+++ Python/Sherlock_valid_string.py
