+++ Python/Triangle Quest.py
