+++ Python/factorial.py
