+++ Python/pangram.py
