+++ Python/pow.py
@@ -1,0 +1,2 @@
+def pow(x, y):
+    return x**y
