+++ Python/prim.py
