+++ Python/second largest.py
