--- pow.py
@@ -1,2 +1,0 @@
-def pow(x, y):
-    return x**y
